Add event listeners to navbar
